requirements.txt
absl-py==0.15.0
alembic==1.7.7
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037829118/work
astunparse==1.6.3
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831763871/work
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.2.1
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.11.0
databricks-cli==0.17.8
dataclasses==0.8
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
docker==5.0.3
eli5==0.13.0
entrypoints==0.3
Flask==2.0.3
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
greenlet==2.0.2
grpcio==1.34.1
h5py==3.1.0
idna==3.7
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916885413/work
importlib-resources==5.4.0
ipykernel @ file:///C:/ci/ipykernel_1596206775157/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==2.0.1
jedi @ file:///C:/ci/jedi_1606914470086/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib==1.1.1
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-core @ file:///C:/ci/jupyter_core_1633402348205/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Mako==1.1.6
Markdown==3.3.7
MarkupSafe @ file:///C:/ci/markupsafe_1621528313524/work
matplotlib==3.3.4
mistune==0.8.4
mlflow==1.23.1
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601896775581/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///C:/ci/notebook_1629205950149/work
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.1.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso==0.7.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
prometheus-client @ file:///opt/conda/conda-bld/prometheus_client_1643788673601/work
prometheus-flask-exporter==0.23.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.6
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT==2.4.0
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent @ file:///C:/ci/pyrsistent_1600141799440/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2024.1
pywin32==227
pywinpty==0.5.7
PyYAML==6.0.1
pyzmq @ file:///C:/ci/pyzmq_1628276569449/work
querystring-parser==1.2.4
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
selenium==3.141.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six==1.15.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.52
sqlparse==0.4.4
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==3.1.0
tornado @ file:///C:/ci/tornado_1606942379977/work
traitlets @ file:///C:/ci/traitlets_1632759765830/work
typing-extensions==3.7.4.3
urllib3==1.26.18
waitress==2.0.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
